Move EglImage to platform/linux.rs; un-gate portable GL modules (A5a)
EglImage is the Linux GlPlatform::Import — its Drop already runs the
Linux-only display lifecycle lock, so it belongs beside the rest of the
Linux platform code. With it gone, cache.rs, resources.rs, and
shaders.rs contain no platform types (gls bindings are runtime-loaded
function pointers and compile everywhere), so their cfg(linux) mounts
and the function! macro's gate are removed: the portable half of the
engine now compiles on macOS as dead code ahead of the engine itself.

No behavior change on either platform. 3 cfg lanes clean; native suite
261/0.
